solve gmpz issues

+++ src/goldilocks_base_field_tools.hpp
@@ -39,11 +39,9 @@ inline Goldilocks::Element Goldilocks::fromS64(int64_t in1)
 
 inline void Goldilocks::fromS64(Element &result, int64_t in1)
 {
-    //#if 0
     uint64_t aux;
     (in1 < 0) ? aux = static_cast<uint64_t>(in1) + GOLDILOCKS_PRIME : aux = static_cast<uint64_t>(in1);
     result.fe = aux;
-    //#endif
 }
 
 
@@ -71,11 +69,9 @@ inline Goldilocks::Element Goldilocks::fromString(const std::string &in1, int ra
 
 inline void Goldilocks::fromString(Element &result, const std::string &in1, int radix)
 {
-//#if 0
     mpz_class aux(in1, radix);
-    aux = (aux + (uint64_t)GOLDILOCKS_PRIME) % (uint64_t)GOLDILOCKS_PRIME;
+    aux = (aux + GOLDILOCKS_PRIME) % GOLDILOCKS_PRIME;
     result.fe = aux.get_ui();
-//#endif
 
 };
 
@@ -88,10 +84,8 @@ inline Goldilocks::Element Goldilocks::fromScalar(const mpz_class &scalar)
 
 inline void Goldilocks::fromScalar(Element &result, const mpz_class &scalar)
 {
-//#if 0
-    mpz_class aux = (scalar + (uint64_t)GOLDILOCKS_PRIME) % (uint64_t)GOLDILOCKS_PRIME;
+    mpz_class aux = (scalar + GOLDILOCKS_PRIME) % GOLDILOCKS_PRIME;
     result.fe = aux.get_ui();
-//#endif
 };
 
 inline uint64_t Goldilocks::toU64(const Element &in1)
@@ -118,29 +112,27 @@ inline int64_t Goldilocks::toS64(const Element &in1)
 inline void Goldilocks::toS64(int64_t &result, const Element &in1)
 {
 
-//#if 0
-    mpz_class out = Goldilocks::toU64(in1);
+    mpz_class out = (unsigned long int) Goldilocks::toU64(in1);
 
-    mpz_class maxInt(((uint64_t)GOLDILOCKS_PRIME - 1) / 2);
+    mpz_class maxInt((GOLDILOCKS_PRIME - 1) / 2);
 
     if (out > maxInt)
     {
-        mpz_class onegative = (uint64_t)GOLDILOCKS_PRIME - out;
+        mpz_class onegative = GOLDILOCKS_PRIME - out;
         result = -onegative.get_si();
     }
     else
     {
         result = out.get_si();
     }
-//#endif
+
 }
 
 /* Converts a field element into a signed 32bits integer */
 /* Precondition:  Goldilocks::Element < 2^31 */
 inline bool Goldilocks::toS32(int32_t &result, const Element &in1)
 {
-//#if 0
-    mpz_class out = mpz_class(Goldilocks::toU64(in1));
+    mpz_class out = mpz_class((unsigned long int) Goldilocks::toU64(in1));
 
     mpz_class maxInt(0x7FFFFFFF);
     mpz_class minInt = mpz_class(0xFFFFFFFF00000001)- mpz_class(0x80000000);
@@ -162,7 +154,6 @@ inline bool Goldilocks::toS32(int32_t &result, const Element &in1)
     {
         result = out.get_si();
     }
-//#endif
     return true;
 }
